+++ flake8_raise.py
@@ -12,6 +12,7 @@ class RaiseStatementChecker:
     text = {
         "R100": "R100 raise in except handler without from",
         "R101": "R101 use bare raise in except handler",
+        "R102": "R102 unnecessary parentheses on raised exception",
     }
 
     def __init__(self, tree, filename):
@@ -27,17 +28,29 @@ class RaiseStatementChecker:
         elif isinstance(node, ast.ExceptHandler):
             self.handlers[-1].append(node.name)
         elif isinstance(node, ast.Raise):
-            if self.handlers[-1] and node.exc is not None:
-                if node.cause is None and not (
-                    isinstance(node.exc, ast.Name) and node.exc.id in self.handlers[-1]
-                ):
-                    yield node.lineno, node.col_offset, self.text["R100"], type(self)
+            if node.exc is not None:
+                if self.handlers[-1]:
+                    if node.cause is None and not (
+                        isinstance(node.exc, ast.Name)
+                        and node.exc.id in self.handlers[-1]
+                    ):
+                        yield node.lineno, node.col_offset, self.text["R100"], type(
+                            self
+                        )
 
-                if (
-                    isinstance(node.exc, ast.Name)
-                    and node.exc.id == self.handlers[-1][-1]
+                    if (
+                        isinstance(node.exc, ast.Name)
+                        and node.exc.id == self.handlers[-1][-1]
+                    ):
+                        yield node.lineno, node.col_offset, self.text["R101"], type(
+                            self
+                        )
+                elif (
+                    isinstance(node.exc, ast.Call)
+                    and not node.exc.args
+                    and not node.exc.keywords
                 ):
-                    yield node.lineno, node.col_offset, self.text["R101"], type(self)
+                    yield node.lineno, node.col_offset, self.text["R102"], type(self)
 
         for child in ast.iter_child_nodes(node):
             yield from self.walk(child)