+++ tests/data/raise_with_attribute_without_parentheses.py
@@ -1,0 +1,2 @@
+var = None
+raise var.exception