+++ tests/data/raise_with_empty_parentheses_noqa.py
@@ -1,0 +1,5 @@
+def get_exception_cls():
+    return ZeroDivisionError
+
+
+raise get_exception_cls()  # noqa: R102